Champion
Added render metrics

+++ OgreMain/include/OgreCamera.h
@@ -75,16 +75,16 @@ namespace Ogre {
         method SceneNode::attachObject. By default the camera is attached to
         the root scene node on creation.
         When this is done the Camera will combine it's own
-        position/orientation settings with it's parent SceneNode. 
+        position/orientation settings with it's parent SceneNode.
         This is also useful for implementing more complex Camera / object
         relationships i.e. having a camera attached to a world object.
     */
     class _OgreExport Camera : public Frustum
     {
     public:
-        /** Listener interface so you can be notified of Camera events. 
+        /** Listener interface so you can be notified of Camera events.
         */
-        class _OgreExport Listener 
+        class _OgreExport Listener
         {
         public:
             Listener() {}
@@ -127,10 +127,10 @@ namespace Ogre {
         Vector3 mYawFixedAxis;
 
         /// Stored number of visible faces in the last render
-        unsigned int mVisFacesLastRender;
+        size_t mVisFacesLastRender;
 
         /// Stored number of visible batches in the last render
-        unsigned int mVisBatchesLastRender;
+        size_t mVisBatchesLastRender;
 
         /// Shared class-level name for Movable type
         static String msMovableType;
@@ -145,20 +145,20 @@ namespace Ogre {
         /// Inverted scene LOD factor, can be used by Renderables to adjust their LOD
         Real mSceneLodFactorInv;
 
-        /** Viewing window. 
+        /** Viewing window.
         @remarks
         Generalize camera class for the case, when viewing frustum doesn't cover all viewport.
         */
         Real mWLeft, mWTop, mWRight, mWBottom;
         /// Is viewing window used.
         bool mWindowSet;
-        /// Windowed viewport clip planes 
+        /// Windowed viewport clip planes
         mutable vector<Plane>::type mWindowClipPlanes;
         /// Was viewing window changed.
         mutable bool mRecalcWindow;
         /// The last viewport to be added using this camera
         Viewport* mLastViewport;
-        /** Whether aspect ratio will automatically be recalculated 
+        /** Whether aspect ratio will automatically be recalculated
             when a viewport changes its size
         */
         bool mAutoAspectRatio;
@@ -168,7 +168,7 @@ namespace Ogre {
         bool mUseRenderingDistance;
         /// Camera to use for LOD calculation
         const Camera* mLodCamera;
-        
+
         /// Whether or not the minimum display size of objects should take effect for this camera
         bool mUseMinPixelSize;
         /// @see Camera::getPixelDisplayRatio
@@ -302,21 +302,21 @@ namespace Ogre {
         */
         void rotate(const Quaternion& q);
 
-        /** Tells the camera whether to yaw around it's own local Y axis or a 
+        /** Tells the camera whether to yaw around it's own local Y axis or a
             fixed axis of choice.
         @remarks
             This method allows you to change the yaw behaviour of the camera
-            - by default, the camera yaws around a fixed Y axis. This is 
-            often what you want - for example if you're making a first-person 
-            shooter, you really don't want the yaw axis to reflect the local 
-            camera Y, because this would mean a different yaw axis if the 
+            - by default, the camera yaws around a fixed Y axis. This is
+            often what you want - for example if you're making a first-person
+            shooter, you really don't want the yaw axis to reflect the local
+            camera Y, because this would mean a different yaw axis if the
             player is looking upwards rather than when they are looking
-            straight ahead. You can change this behaviour by calling this 
+            straight ahead. You can change this behaviour by calling this
             method, which you will want to do if you are making a completely
-            free camera like the kind used in a flight simulator. 
+            free camera like the kind used in a flight simulator.
         @param useFixed
-            If @c true, the axis passed in the second parameter will 
-            always be the yaw axis no matter what the camera orientation. 
+            If @c true, the axis passed in the second parameter will
+            always be the yaw axis no matter what the camera orientation.
             If false, the camera yaws around the local Y.
         @param fixedAxis
             The axis to use if the first parameter is true.
@@ -367,19 +367,19 @@ namespace Ogre {
 
         /** Internal method to notify camera of the visible faces in the last render.
         */
-        void _notifyRenderedFaces(unsigned int numfaces);
+        void _notifyRenderedFaces(size_t numfaces);
 
         /** Internal method to notify camera of the visible batches in the last render.
         */
-        void _notifyRenderedBatches(unsigned int numbatches);
+        void _notifyRenderedBatches(size_t numbatches);
 
         /** Internal method to retrieve the number of visible faces in the last render.
         */
-        unsigned int _getNumRenderedFaces(void) const;
+        size_t _getNumRenderedFaces(void) const;
 
         /** Internal method to retrieve the number of visible batches in the last render.
         */
-        unsigned int _getNumRenderedBatches(void) const;
+        size_t _getNumRenderedBatches(void) const;
 
         /** Gets the derived orientation of the camera, including any
             rotation inherited from a node attachment and reflection matrix. */
@@ -429,14 +429,14 @@ namespace Ogre {
         /** Enables / disables automatic tracking of a SceneNode.
         @remarks
             If you enable auto-tracking, this Camera will automatically rotate to
-            look at the target SceneNode every frame, no matter how 
+            look at the target SceneNode every frame, no matter how
             it or SceneNode move. This is handy if you want a Camera to be focused on a
-            single object or group of objects. Note that by default the Camera looks at the 
+            single object or group of objects. Note that by default the Camera looks at the
             origin of the SceneNode, if you want to tweak this, e.g. if the object which is
             attached to this target node is quite big and you want to point the camera at
-            a specific point on it, provide a vector in the 'offset' parameter and the 
+            a specific point on it, provide a vector in the 'offset' parameter and the
             camera's target point will be adjusted.
-        @param enabled If true, the Camera will track the SceneNode supplied as the next 
+        @param enabled If true, the Camera will track the SceneNode supplied as the next
             parameter (cannot be null). If false the camera will cease tracking and will
             remain in it's current orientation.
         @param target Pointer to the SceneNode which this Camera will track. Make sure you don't
@@ -445,17 +445,17 @@ namespace Ogre {
         @param offset If supplied, the camera targets this point in local space of the target node
             instead of the origin of the target node. Good for fine tuning the look at point.
         */
-        void setAutoTracking(bool enabled, SceneNode* const target = 0, 
+        void setAutoTracking(bool enabled, SceneNode* const target = 0,
             const Vector3& offset = Vector3::ZERO);
 
 
         /** Sets the level-of-detail factor for this Camera.
         @remarks
-            This method can be used to influence the overall level of detail of the scenes 
+            This method can be used to influence the overall level of detail of the scenes
             rendered using this camera. Various elements of the scene have level-of-detail
-            reductions to improve rendering speed at distance; this method allows you 
+            reductions to improve rendering speed at distance; this method allows you
             to hint to those elements that you would like to adjust the level of detail that
-            they would normally use (up or down). 
+            they would normally use (up or down).
         @par
             The most common use for this method is to reduce the overall level of detail used
             for a secondary camera used for sub viewports like rear-view mirrors etc.
@@ -466,31 +466,31 @@ namespace Ogre {
         */
         void setLodBias(Real factor = 1.0);
 
-        /** Returns the level-of-detail bias factor currently applied to this camera. 
+        /** Returns the level-of-detail bias factor currently applied to this camera.
         @remarks
             See Camera::setLodBias for more details.
         */
         Real getLodBias(void) const;
 
         /** Set a pointer to the camera which should be used to determine
-            LOD settings. 
+            LOD settings.
         @remarks
             Sometimes you don't want the LOD of a render to be based on the camera
             that's doing the rendering, you want it to be based on a different
-            camera. A good example is when rendering shadow maps, since they will 
+            camera. A good example is when rendering shadow maps, since they will
             be viewed from the perspective of another camera. Therefore this method
             lets you associate a different camera instance to use to determine the LOD.
         @par
-            To revert the camera to determining LOD based on itself, call this method with 
-            a pointer to itself. 
+            To revert the camera to determining LOD based on itself, call this method with
+            a pointer to itself.
         */
         virtual void setLodCamera(const Camera* lodCam);
 
-        /** Get a pointer to the camera which should be used to determine 
-            LOD settings. 
+        /** Get a pointer to the camera which should be used to determine
+            LOD settings.
         @remarks
             If setLodCamera hasn't been called with a different camera, this
-            method will return 'this'. 
+            method will return 'this'.
         */
         virtual const Camera* getLodCamera() const;
 
@@ -501,38 +501,38 @@ namespace Ogre {
         */
         Ray getCameraToViewportRay(Real screenx, Real screeny) const;
         /** Gets a world space ray as cast from the camera through a viewport position.
-        @param screenx, screeny The x and y position at which the ray should intersect the viewport, 
+        @param screenx, screeny The x and y position at which the ray should intersect the viewport,
             in normalised screen coordinates [0,1]
         @param outRay Ray instance to populate with result
         */
         void getCameraToViewportRay(Real screenx, Real screeny, Ray* outRay) const;
 
         /** Gets a world-space list of planes enclosing a volume based on a viewport
-            rectangle. 
+            rectangle.
         @remarks
-            Can be useful for populating a PlaneBoundedVolumeListSceneQuery, e.g. 
-            for a rubber-band selection. 
+            Can be useful for populating a PlaneBoundedVolumeListSceneQuery, e.g.
+            for a rubber-band selection.
         @param screenLeft, screenTop, screenRight, screenBottom The bounds of the
             on-screen rectangle, expressed in normalised screen coordinates [0,1]
-        @param includeFarPlane If true, the volume is truncated by the camera far plane, 
+        @param includeFarPlane If true, the volume is truncated by the camera far plane,
             by default it is left open-ended
         */
-        PlaneBoundedVolume getCameraToViewportBoxVolume(Real screenLeft, 
+        PlaneBoundedVolume getCameraToViewportBoxVolume(Real screenLeft,
             Real screenTop, Real screenRight, Real screenBottom, bool includeFarPlane = false);
 
         /** Gets a world-space list of planes enclosing a volume based on a viewport
-            rectangle. 
+            rectangle.
         @remarks
-            Can be useful for populating a PlaneBoundedVolumeListSceneQuery, e.g. 
-            for a rubber-band selection. 
+            Can be useful for populating a PlaneBoundedVolumeListSceneQuery, e.g.
+            for a rubber-band selection.
         @param screenLeft, screenTop, screenRight, screenBottom The bounds of the
             on-screen rectangle, expressed in normalised screen coordinates [0,1]
         @param outVolume The plane list to populate with the result
-        @param includeFarPlane If true, the volume is truncated by the camera far plane, 
+        @param includeFarPlane If true, the volume is truncated by the camera far plane,
             by default it is left open-ended
         */
-        void getCameraToViewportBoxVolume(Real screenLeft, 
-            Real screenTop, Real screenRight, Real screenBottom, 
+        void getCameraToViewportBoxVolume(Real screenLeft,
+            Real screenTop, Real screenRight, Real screenBottom,
             PlaneBoundedVolume* outVolume, bool includeFarPlane = false);
 
         /** Internal method for OGRE to use for LOD calculations. */
@@ -546,7 +546,7 @@ namespace Ogre {
         /** Sets the viewing window inside of viewport.
         @remarks
             This method can be used to set a subset of the viewport as the rendering
-            target. 
+            target.
         @param left Relative to Viewport - 0 corresponds to left edge, 1 - to right edge (default - 0).
         @param top Relative to Viewport - 0 corresponds to top edge, 1 - to bottom edge (default - 0).
         @param right Relative to Viewport - 0 corresponds to left edge, 1 - to right edge (default - 1).
@@ -564,8 +564,8 @@ namespace Ogre {
         SceneNode* getAutoTrackTarget(void) const { return mAutoTrackTarget; }
         /** Get the auto tracking offset for this camera, if it is auto tracking. */
         const Vector3& getAutoTrackOffset(void) const { return mAutoTrackOffset; }
-        
-        /** Get the last viewport which was attached to this camera. 
+
+        /** Get the last viewport which was attached to this camera.
         @note This is not guaranteed to be the only viewport which is
             using this camera, just the last once which was created referring
             to it.
@@ -574,13 +574,13 @@ namespace Ogre {
         /** Notifies this camera that a viewport is using it.*/
         void _notifyViewport(Viewport* viewport) {mLastViewport = viewport;}
 
-        /** If set to true a viewport that owns this frustum will be able to 
+        /** If set to true a viewport that owns this frustum will be able to
             recalculate the aspect ratio whenever the frustum is resized.
         @remarks
-            You should set this to true only if the frustum / camera is used by 
-            one viewport at the same time. Otherwise the aspect ratio for other 
+            You should set this to true only if the frustum / camera is used by
+            one viewport at the same time. Otherwise the aspect ratio for other
             viewports may be wrong.
-        */    
+        */
         void setAutoAspectRatio(bool autoratio);
 
         /** Retrieves if AutoAspectRatio is currently set or not
@@ -592,7 +592,7 @@ namespace Ogre {
             By calling this method, you can tell the camera to perform culling
             against a different frustum to it's own. This is mostly useful for
             debug cameras that allow you to show the culling behaviour of another
-            camera, or a manual frustum instance. 
+            camera, or a manual frustum instance.
         @param frustum Pointer to a frustum to use; this can either be a manual
             Frustum instance (which you can attach to scene nodes like any other
             MovableObject), or another camera. If you pass 0 to this method it
@@ -619,7 +619,7 @@ namespace Ogre {
         /// @copydoc Frustum::getFrustumPlane
         const Plane& getFrustumPlane( unsigned short plane ) const;
         /// @copydoc Frustum::projectSphere
-        bool projectSphere(const Sphere& sphere, 
+        bool projectSphere(const Sphere& sphere,
             Real* left, Real* top, Real* right, Real* bottom) const;
         /// @copydoc Frustum::getNearClipDistance
         Real getNearClipDistance(void) const;
@@ -628,10 +628,10 @@ namespace Ogre {
         /// @copydoc Frustum::getViewMatrix
         const Matrix4& getViewMatrix(void) const;
         /** Specialised version of getViewMatrix allowing caller to differentiate
-            whether the custom culling frustum should be allowed or not. 
+            whether the custom culling frustum should be allowed or not.
         @remarks
-            The default behaviour of the standard getViewMatrix is to delegate to 
-            the alternate culling frustum, if it is set. This is expected when 
+            The default behaviour of the standard getViewMatrix is to delegate to
+            the alternate culling frustum, if it is set. This is expected when
             performing CPU calculations, but the final rendering must be performed
             using the real view matrix in order to display the correct debug view.
         */
@@ -647,10 +647,10 @@ namespace Ogre {
         */
         virtual bool getUseRenderingDistance(void) const { return mUseRenderingDistance; }
 
-        /** Synchronise core camera settings with another. 
+        /** Synchronise core camera settings with another.
         @remarks
-            Copies the position, orientation, clip distances, projection type, 
-            FOV, focal length and aspect ratio from another camera. Other settings like query flags, 
+            Copies the position, orientation, clip distances, projection type,
+            FOV, focal length and aspect ratio from another camera. Other settings like query flags,
             reflection etc are preserved.
         */
         virtual void synchroniseBaseSettingsWith(const Camera* cam);
@@ -665,7 +665,7 @@ namespace Ogre {
             be rendered.
         */
         void setUseMinPixelSize(bool enable) { mUseMinPixelSize = enable; }
-        /** Returns whether to use min display size calculations 
+        /** Returns whether to use min display size calculations
         @see Camera::setUseMinDisplaySize
         */
         bool getUseMinPixelSize() const { return mUseMinPixelSize; }
@@ -673,7 +673,7 @@ namespace Ogre {
         /** Returns an estimated ratio between a pixel and the display area it represents.
             For orthographic cameras this function returns the amount of meters covered by
             a single pixel along the vertical axis. For perspective cameras the value
-            returned is the amount of meters covered by a single pixel per meter distance 
+            returned is the amount of meters covered by a single pixel per meter distance
             from the camera.
         @note
             This parameter is calculated just before the camera is rendered
@@ -702,7 +702,7 @@ namespace Ogre {
 
         /// Returns true if the asked render queue has been rendered. False otherwise
         bool isRenderedRq( size_t rqId ) const          { return mRenderedRqs[rqId]; }
-        
+
     };
     /** @} */
     /** @} */
